feat: update

+++ src/app/services/event.service.ts
@@ -4,9 +4,7 @@ import { Event } from '../interfaces/event';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { GroupService } from './group.service';
 import { map, switchMap } from 'rxjs/operators';
-import { Observable, combineLatest, of } from 'rxjs';
-import { firestore } from 'firebase';
-import { Group } from '../interfaces/group';
+import { Observable, combineLatest, of, ObservableLike } from 'rxjs';
 import { Id } from '../interfaces/id';
 @Injectable({
   providedIn: 'root',
@@ -24,12 +22,9 @@ export class EventService {
       .doc(`events/${id}`)
       .set(event)
       .then(() =>
-        this.db.doc(`groups/${event.groupid}/eventIds`).set({ id: event.id })
-      )
-      .then(() =>
-        this.snackbar.open('Successfully created the event', null, {
-          duration: 2000,
-        })
+        this.db
+          .doc(`groups/${event.groupid}/eventIds/${event.id}`)
+          .set({ id: event.id })
       );
   }
 
@@ -39,35 +34,45 @@ export class EventService {
 
   getMyEvents(uid: string): Observable<Event[]> {
     return this.db
-      .collection<Id>(`users/${uid}/eventIds`)
+      .collection<Id>(`users/${uid}/groupIds`)
       .valueChanges()
       .pipe(
-        map((eventIds: Id[]) => {
-          if (eventIds.length) {
-            const result: Observable<Event>[] = [];
-            eventIds.forEach((eventId: Id) => {
+        map((Ids: Id[]) => {
+          console.log(Ids);
+          const groupIds: string[] = [];
+          Ids.map((id: Id) => {
+            groupIds.push(id.id);
+          });
+          return groupIds;
+        }),
+        switchMap((groupIds: string[]) => {
+          const eventIdsList: Observable<Id[]>[] = [];
+          groupIds.map((groupId: string) => {
+            const eventIds: Observable<Id[]> = this.db
+              .collection<Id>(`groups/${groupId}/eventIds`)
+              .valueChanges();
+            eventIdsList.push(eventIds);
+          });
+          return combineLatest(eventIdsList);
+        }),
+        switchMap((eventIdsList: Id[][]) => {
+          const events: Observable<Event>[] = [];
+          eventIdsList.map((eventIds: Id[]) => {
+            eventIds.map((eventId: Id) => {
               const event: Observable<Event> = this.db
                 .doc<Event>(`events/${eventId.id}`)
                 .valueChanges();
-              result.push(event);
+              events.push(event);
             });
-            return combineLatest(result);
-          } else {
-            return of([]);
-          }
+          });
+          console.log(events);
+          return combineLatest(events);
         })
       );
   }
 
   async updateEvent(uid: string, event: Omit<Event, 'createrId'>) {
-    await this.db
-      .doc(`events/${event.id}`)
-      .set(event, { merge: true })
-      .then(() =>
-        this.snackbar.open('Successfully updated the event', null, {
-          duration: 2000,
-        })
-      );
+    await this.db.doc(`events/${event.id}`).set(event, { merge: true });
   }
 
   async deleteEvent(eventId: string, groupId: string) {
@@ -80,7 +85,7 @@ export class EventService {
           .collection<Id>(`groups/${groupId}/attendingMemberIds`)
           .valueChanges()
           .subscribe((attendingMemberIds: Id[]) => {
-            attendingMemberIds.forEach((attendingMemberId: Id) => {
+            attendingMemberIds.map((attendingMemberId: Id) => {
               this.db
                 .doc(
                   `groups/${groupId}/attendingMemberIds/${attendingMemberId.id}`
@@ -94,7 +99,7 @@ export class EventService {
           .collection(`groups/${groupId}/waitingJoinningMemberIds`)
           .valueChanges()
           .subscribe((waitingJoinningMemberIds: Id[]) => {
-            waitingJoinningMemberIds.forEach((waitingJoinningMemberId: Id) => {
+            waitingJoinningMemberIds.map((waitingJoinningMemberId: Id) => {
               this.db
                 .doc(
                   `groups/${groupId}/attendingMemberIds/${waitingJoinningMemberId.id}`
@@ -108,7 +113,7 @@ export class EventService {
           .collection(`groups/${groupId}/waitingPayingMemberIds`)
           .valueChanges()
           .subscribe((waitingPayingMemberIds: Id[]) => {
-            waitingPayingMemberIds.forEach((waitingPayingMemberId: Id) => {
+            waitingPayingMemberIds.map((waitingPayingMemberId: Id) => {
               this.db
                 .doc(
                   `groups/${groupId}/attendingMemberIds/${waitingPayingMemberId.id}`
@@ -117,11 +122,18 @@ export class EventService {
             });
           });
       })
-      .then(() =>
-        this.snackbar.open('Successfully deleted the event', null, {
-          duration: 2000,
-        })
-      );
+      .then(() => {
+        this.db
+          .collection(`events/${eventId}/attendingMemberIds`)
+          .valueChanges()
+          .subscribe((attendingMemberIds: Id[]) => {
+            attendingMemberIds.map((attendingMemberId: Id) => {
+              this.db
+                .doc(`users/${attendingMemberId.id}/eventIds/${eventId}`)
+                .delete();
+            });
+          });
+      });
   }
 
   getSearchableEvents(): Observable<Event[]> {
@@ -130,14 +142,84 @@ export class EventService {
       .valueChanges();
   }
 
+  getAttendingMemberIds(eventId: string): Observable<string[]> {
+    return this.db
+      .collection(`events/${eventId}/attendingMemberIds`)
+      .valueChanges()
+      .pipe(
+        map((attendingMemberIds: Id[]) => {
+          if (attendingMemberIds.length) {
+            const result: string[] = [];
+            attendingMemberIds.map((attendingMemberId: Id) => {
+              result.push(attendingMemberId.id);
+            });
+            return result;
+          } else {
+            return [];
+          }
+        })
+      );
+  }
+
+  getWaitingPayingMemberIds(eventId: string): Observable<string[]> {
+    return this.db
+      .collection(`events/${eventId}/waitingPayingMemberIds`)
+      .valueChanges()
+      .pipe(
+        map((waitingPayingMemberIds: Id[]) => {
+          if (waitingPayingMemberIds.length) {
+            const result: string[] = [];
+            waitingPayingMemberIds.map((waitingPayingMemberId: Id) => {
+              result.push(waitingPayingMemberId.id);
+            });
+            return result;
+          } else {
+            return [];
+          }
+        })
+      );
+  }
+
+  getWaitingJoinningMemberIds(eventId: string): Observable<string[]> {
+    return this.db
+      .collection(`events/${eventId}/waitingJoinningMemberIds`)
+      .valueChanges()
+      .pipe(
+        map((waitingJoinningMemberIds: Id[]) => {
+          if (waitingJoinningMemberIds.length) {
+            const result: string[] = [];
+            waitingJoinningMemberIds.map((waitingJoinningMemberId: Id) => {
+              result.push(waitingJoinningMemberId.id);
+            });
+            return result;
+          } else {
+            return [];
+          }
+        })
+      );
+  }
+
   // need to work
   getAttendingEvents(uid: string): Observable<Event[]> {
-    this.db.collection(`events`);
     return this.db
-      .collection<Event>(`events`, (ref) =>
-        ref.where('attendingMemberIds', 'array-contains', uid)
-      )
-      .valueChanges();
+      .collection(`users/${uid}/eventIds`)
+      .valueChanges()
+      .pipe(
+        switchMap((eventIds: Id[]) => {
+          if (eventIds.length) {
+            const result: Observable<Event>[] = [];
+            eventIds.map((eventId: Id) => {
+              const attendingEvent: Observable<Event> = this.db
+                .doc<Event>(`events/${eventId.id}`)
+                .valueChanges();
+              result.push(attendingEvent);
+            });
+            return combineLatest(result);
+          } else {
+            return of([]);
+          }
+        })
+      );
   }
 
   // nothing to attending (pay+public, pay+private, free+public, free+private)
@@ -153,71 +235,68 @@ export class EventService {
   // waitingPayinglist to attending (pay+public, pay+private, free+public, free+private)
   async payToAttendEvent(uid: string, eventId: string) {
     await this.db
-      .doc(`events/${eventId}/'attendingMemberIds'/${uid}`)
+      .doc(`events/${eventId}/attendingMemberIds/${uid}`)
       .set({ id: uid })
       .then(() => {
-        this.db
-          .doc(`events/${eventId}/'waitingPayingMemberIds'/${uid}`)
-          .delete();
+        this.db.doc(`users/${uid}/eventIds/${eventId}`).set({ id: eventId });
       })
       .then(() => {
-        this.db.doc(`users/${uid}/eventIds/${eventId}`).set({ id: eventId });
+        this.db.doc(`events/${eventId}/waitingPayingMemberIds/${uid}`).delete();
       });
   }
 
   // attending to nothing (pay+public, pay+private, free+public, free+private)
-  async leaveEvent(uid: string, eventid: string) {
+  async leaveEvent(uid: string, eventId: string) {
     await this.db
-      .doc(`events/${eventid}`)
-      .update({ attendingMemberIds: firestore.FieldValue.arrayRemove(uid) });
+      .doc(`events/${eventId}/attendingMemberIds/${uid}`)
+      .delete()
+      .then(() => {
+        this.db.doc(`users/${uid}/eventIds/${eventId}`).delete();
+      });
   }
 
   // nothing to waitingJoinning (pay+private, free+private)
   async joinWaitingJoinningList(uid: string, eventId: string) {
-    await this.db.doc(`events/${eventId}`).update({
-      waitingJoinningMemberIds: firestore.FieldValue.arrayUnion(uid),
-    });
+    await this.db
+      .doc(`events/${eventId}/waitingJoinningMemberIds/${uid}`)
+      .set({ id: uid });
   }
 
   // waitingJoinning to waitingPaying (pay+private)
   async joinWaitingPayingList(uid: string, eventId: string) {
     await this.db
-      .doc(`events/${eventId}`)
-      .update({
-        waitingJoinningMemberIds: firestore.FieldValue.arrayRemove(uid),
-      })
+      .doc(`events/${eventId}/waitingJoinningMemberIds/${uid}`)
+      .delete()
       .then(() => {
-        this.db.doc(`events/${eventId}`).update({
-          waitingPayingMemberIds: firestore.FieldValue.arrayUnion(uid),
-        });
+        this.db
+          .doc(`events/${eventId}/waitingPayingMemberIds/${uid}`)
+          .set({ id: uid });
       });
   }
 
   // waitingJoinning to nothing (pay+private, free+private)
   async removeWaitingJoinningMember(uid: string, eventId: string) {
-    await this.db.doc(`events/${eventId}`).update({
-      waitingJoinningMemberIds: firestore.FieldValue.arrayRemove(uid),
-    });
+    await this.db
+      .doc(`events/${eventId}/waitingJoinningMemberIds/${uid}`)
+      .delete();
   }
 
   // waitingPaying to nothing (pay+private)
   async removeWaitingPayingMember(uid: string, eventId: string) {
-    await this.db.doc(`events/${eventId}`).update({
-      waitingPayingMemberIds: firestore.FieldValue.arrayRemove(uid),
-    });
+    await this.db
+      .doc(`events/${eventId}/waitingPayingMemberIds/${uid}`)
+      .delete();
   }
 
   // waitingJoinning to attending (free+private)
   async waitingJoinningMemberToAttendingMember(uid: string, eventId: string) {
     await this.db
-      .doc(`events/${eventId}`)
-      .update({
-        waitingJoinningMemberIds: firestore.FieldValue.arrayRemove(uid),
-      })
+      .doc(`events/${eventId}/waitingJoinningMemberIds/${uid}`)
+      .delete()
       .then(() => {
         this.db
-          .doc(`events/${eventId}`)
-          .update({ attendingMemberIds: firestore.FieldValue.arrayUnion(uid) });
+          .doc(`events/${eventId}/attendingMemberIds/${uid}`)
+          .set({ id: uid });
       });
   }
 }
